requirements.txt
faker==40.4.0
pytest==9.0.2
pytest-html==4.2.0
python-dotenv==1.2.1
playwright==1.58.0